Update specs, reduce duplication around testing SFTP client
