+++ spec/importers/x2_importer_spec.rb
@@ -11,23 +11,6 @@ RSpec.describe do
     end
   }
 
-  describe '#connect_and_import' do
-    let(:x2_importer) { x2_student_import_class.new }
-    context 'sftp information present' do
-      it 'does not raise an error' do
-        allow(ENV).to receive(:[]).with('SIS_SFTP_HOST').and_return "sftp-site@site.com"
-        allow(ENV).to receive(:[]).with('SIS_SFTP_USER').and_return "sftp-user"
-        allow(ENV).to receive(:[]).with('SIS_SFTP_KEY').and_return "sftp-key"
-        allow(Net::SFTP).to receive_messages(start: 'connection established')
-        expect { x2_importer.connect_and_import }.not_to raise_error
-      end
-    end
-    context 'sftp information missing' do
-      it 'raises an error' do
-        expect { x2_importer.connect_and_import }.to raise_error "SFTP information missing"
-      end
-    end
-  end
   describe '#import' do
     context 'students' do
       let(:x2_importer) { x2_student_import_class.new }
